Move header component to separate file
- Extract header code from App component into Header.tsx
- Update App component to use the new Header component
- Improve code organization and readability
